requirements.txt
einops==0.6.1
einops-exts==0.0.4
huggingface-hub==0.16.4
nibabel==5.1.0
nmslib==2.1.1
opencv-python==4.8.0.76
pandas==2.0.3
Pillow==9.4.0
pytz==2023.3
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.11.2
# scispacy==0.5.2
sentencepiece==0.1.99
SimpleITK==2.2.1
# spacy==3.6.1
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.5
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.28.1
